+++ src/main/java/com/travel/travel_project/domain/common/CommonDTO.java
@@ -19,7 +19,7 @@ import javax.validation.constraints.NotEmpty;
 public class CommonDTO extends NewCommonDTO {
 
     @ApiModelProperty(required = true, value = "rnum", hidden = true, example = "1")
-    private Integer rnum;
+    private Integer rowNum;
 
     @ApiModelProperty(required = true, value = "idx", hidden = true, example = "1")
     private Long idx;
